refactor: rename `components/NewDataExplorer` to `components/FluxQueryBuilder`
